+++ example/example.dart
@@ -1,23 +1,21 @@
-import 'package:diffie_hellman/src/engines/dh_pkcs3_engine.dart';
-import 'package:diffie_hellman/src/models/dh_key_pair.dart';
-import 'package:flutter/cupertino.dart';
+import 'package:diffie_hellman/diffie_hellman.dart';
 
 void main() {
-  DhPkcs3Engine dhEngine = DhPkcs3Engine.fromGroup(5);
-  DhPkcs3Engine otherDhEngine = DhPkcs3Engine.fromGroup(5);
+  DhPkcs3Engine dhEngine = DhPkcs3Engine.fromGroup(DhGroup.g5);
+  DhPkcs3Engine otherDhEngine = DhPkcs3Engine.fromGroup(DhGroup.g5);
 
   DhKeyPair keyPair = dhEngine.generateKeyPair();
   DhKeyPair otherKeyPair = otherDhEngine.generateKeyPair();
 
-  debugPrint('Public Key: ${keyPair.publicKey}');
-  debugPrint('Private Key: ${keyPair.privateKey}');
-  debugPrint('Other public Key: ${otherKeyPair.publicKey}');
-  debugPrint('Other private Key: ${otherKeyPair.privateKey}');
+  print('Public Key: ${keyPair.publicKey}');
+  print('Private Key: ${keyPair.privateKey}');
+  print('Other public Key: ${otherKeyPair.publicKey}');
+  print('Other private Key: ${otherKeyPair.privateKey}');
 
-  debugPrint(
+  print(
     'Secret Key: ${dhEngine.computeSecretKey(otherKeyPair.publicKey)}',
   );
-  debugPrint(
+  print(
     'Other secret Key: ${otherDhEngine.computeSecretKey(keyPair.publicKey)}',
   );
 }