+++ lib/diffie_hellman.dart
@@ -1,5 +1,6 @@
 library diffie_hellman;
 
 export 'src/engines/dh_pkcs3_engine.dart';
-export 'src/models/dh_parameter_spec.dart';
-export 'src/models/dh_key_pair.dart';
+export 'src/engines/models/dh_key_pair.dart';
+export 'src/spec/dh_group.dart';
+export 'src/spec/dh_parameter_spec.dart';