+++ CHANGELOG.md
@@ -12,3 +12,9 @@
 
 ## 0.1.2
 * Changed private utility method `toBigInt()`
+
+## 1.0.0
+* Bumped minimal Dart SDK version to 2.17.0
+* Added DhGroup enum instead of int groupId
+* Added DhPkcs3Engine.fromKeyPair() factory
+* Allowed key pair regeneration in DhPkcs3Engine